requirements.txt
tensorflow==2.17.1
deepface==0.0.93
streamlit==1.30.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
numpy>=1.25.0
pandas==2.1.4
Pillow==10.1.0
mtcnn==0.1.1
bcrypt==4.1.2
PyYAML==6.0.1

